Update the fix for BatchNorm, still use cuddn API cudnnBatchNormalizationForwardInference. The root case is, for half type, should use alpha, beta, scale, B, mean, var with float type.

+++ onnxruntime/core/providers/cuda/cudnn_common.h
@@ -49,8 +49,8 @@ struct Consts {
 
 template <>
 struct Consts<half> {
-  static const half Zero;
-  static const half One;
+  static const float Zero;
+  static const float One;
 };
 
 inline double ClampCudnnBatchNormEpsilon(double epsilon) {
